+++ tests/Telplin.Core.Tests/TypeTests.fs
@@ -458,8 +458,8 @@ module X
 
 type X =
     new: unit -> X
-    member Item: m: int -> string with get
     member Item: m: int -> string with set
+    member Item: m: int -> string with get
 """
 
 [<Test>]
@@ -994,8 +994,8 @@ namespace Sample
 module Inner =
     type Facts =
         new: name1: string * name2: string -> Facts
-        member Name: j: float -> string with get
         member Name: s: string -> float with set
+        member Name: j: float -> string with get
 """
 
 [<Test>]